removed glaze templates.
After thinking about them (at 2 in the morning), I came to the conclusion that they are confusing and useless. The sole purpose of the templates is to make creating several of similar glazes easily. However, in the few cases where this is necessary, copying the first glaze to make another is easier, and would be the first logical jump. Secondly, since they are only used for single glazes and nothing else, they fragment the user experience. This creates an inconsistent feel to the app, running counter to the whole purpose of it - having an easy to use, streamlined, symmetic database. Additionally, fragmenting the user experience confuses the user, and creates a whole other thing for them to learn. It would make the program easier to use, and more importantly, easier to learn to remove them altogether.